+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [] - 2023-11-13
+- Issue 152: Updated FDO components to version 1.1.6
+- Fixed package name typo in Fedora.
+- Changed docker-compose download to pull the minimum allowed version.
+
 ## [] - 2023-07-10
 - Issue 146: Integrated FDO into all-in-1.
 - Removed SDO Support.